update the route for group and participants
